+++ src/lightning_habana/pytorch/strategies/parallel.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import logging
-import os
 from typing import Any, Callable, Dict, List, Optional, Union
 
 import torch