+++ src/lightning_habana/pytorch/strategies/deepspeed.py
@@ -82,8 +82,8 @@ log = logging.getLogger(__name__)
 warning_cache = WarningCache()
 
 _HPU_DEEPSPEED_AVAILABLE = (
-    # HPU deep speed is supported only through this pip install git+https://github.com/HabanaAI/DeepSpeed.git@1.12.0
-    RequirementCache("deepspeed==0.9.4+hpu.synapse.v1.12.0")
+    # HPU deep speed is supported only through this pip install git+https://github.com/HabanaAI/DeepSpeed.git@1.12.1
+    RequirementCache("deepspeed==0.9.4+hpu.synapse.v1.12.1")
 )
 if TYPE_CHECKING and _HPU_DEEPSPEED_AVAILABLE:
     import deepspeed
@@ -295,7 +295,7 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
         if not _HPU_DEEPSPEED_AVAILABLE:
             raise MisconfigurationException(
                 "To use the `HPUDeepSpeedStrategy`, you must have hpu DeepSpeed installed."
-                " Install it by running `pip install git+https://github.com/HabanaAI/DeepSpeed.git@1.12.0`."
+                " Install it by running `pip install git+https://github.com/HabanaAI/DeepSpeed.git@1.12.1`."
             )
 
         super().__init__(