+++ src/lightning_habana/pytorch/strategies/parallel.py
@@ -27,6 +27,7 @@ if module_available("lightning"):
     from lightning.pytorch.plugins.io.wrapper import _WrappingCheckpointIO
     from lightning.pytorch.plugins.precision import PrecisionPlugin
     from lightning.pytorch.strategies.ddp import DDPStrategy
+    from lightning.pytorch.utilities.types import STEP_OUTPUT
 elif module_available("pytorch_lightning"):
     from lightning_fabric.plugins import CheckpointIO, ClusterEnvironment
     from lightning_fabric.utilities.distributed import group as _group
@@ -36,6 +37,7 @@ elif module_available("pytorch_lightning"):
     from pytorch_lightning.plugins.io.wrapper import _WrappingCheckpointIO
     from pytorch_lightning.plugins.precision import PrecisionPlugin
     from pytorch_lightning.strategies.ddp import DDPStrategy
+    from pytorch_lightning.utilities.types import STEP_OUTPUT
 else:
     raise ModuleNotFoundError("You are missing `lightning` or `pytorch-lightning` package, please install it.")
 from torch import Tensor
@@ -138,20 +140,20 @@ class HPUParallelStrategy(DDPStrategy):
         htcore.mark_step()
         return optimizer_output
 
-    def validation_step(self, batch: Any, batch_idx: int) -> Any:
+    def validation_step(self, *args: Any, **kwargs: Any) -> STEP_OUTPUT:
         # Break lazy accumulation of graph after every step
         htcore.mark_step()
-        return super().validation_step(batch, batch_idx)
+        return super().validation_step(*args, **kwargs)
 
-    def test_step(self, batch: Any, batch_idx: int) -> Any:
+    def test_step(self, *args: Any, **kwargs: Any) -> STEP_OUTPUT:
         # Break lazy accumulation of graph after every step
         htcore.mark_step()
-        return super().test_step(batch, batch_idx)
+        return super().test_step(*args, **kwargs)
 
-    def predict_step(self, batch: Any, batch_idx: int) -> Any:
+    def predict_step(self, *args: Any, **kwargs: Any) -> Any:
         # Break lazy accumulation of graph after every step
         htcore.mark_step()
-        return super().predict_step(batch, batch_idx)
+        return super().predict_step(*args, **kwargs)
 
     def reduce(
         self, tensor: Tensor, group: Optional[Any] = None, reduce_op: Optional[Union[ReduceOp, str]] = "mean"