+++ CHANGELOG.md
@@ -5,6 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-MM-DD
+
+
+### Added
+
+-
+
+### Changed
+
+-
+
+### Fixed
+
+- Remove redundant env variable for HPU distributed backend ([#91](https://github.com/Lightning-AI/lightning-Habana/pull/91))
+
+
+### Removed
+
+-
+
+### Deprecated
+
+-
+
+
 ## [1.1.0] - 2023-09-26