+++ CHANGELOG.md
@@ -13,11 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--
+- Changed HPU docker image based on synapse AI release 1.12.0 ([#90](https://github.com/Lightning-AI/lightning-Habana/pull/90))
 
 ### Fixed
 
-- Changed HPU docker image based on synapse AI release 1.12.0 ([#90](https://github.com/Lightning-AI/lightning-Habana/pull/90))
 - Remove redundant env variable for HPU distributed backend ([#91](https://github.com/Lightning-AI/lightning-Habana/pull/91))
 
 ### Removed