+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added tests, examples and documentation for HPUPrecisionPlugin with autocast ([#94](https://github.com/Lightning-AI/lightning-Habana/pull/94))
 - Added test to validate checkpoint resuming with Deepspeed strategy on HPU ([#95](https://github.com/Lightning-AI/lightning-Habana/pull/95))
+- Added support for lightning 2.1 ([#100](https://github.com/Lightning-AI/lightning-Habana/pull/100))
 
 ### Changed